Main

Documentation

type Point = (Float, Float)

winHeight :: Int

Window height

Window configuration

winWidth :: Int

Window width

Window configuration

winXOffset :: Int

Window x axis offset

Window configuration

winYOffset :: Int

Window y axis offset

Window configuration

fps :: Int

Game fps

Window configuration

winBackground :: Color

Window backgroud color (i.e. river color)

window :: Display

Game window description (caption, size, offset)

bankWidth :: Float

Width of bank on every side of the river

riverWidth :: Float

River width

riverWidthHalf :: Float

speedupIndex :: Float

How much faster will be the game with every frame

moveDiff :: Float

How much move the plane in one frame

enemyRadius :: Float

Size of enemy

planeRadius :: Float

Size of player

render :: GameState -> Picture

Render the game

renderGround :: (GameState, [Picture]) -> (GameState, [Picture])

Render the ground around the river

renderPlane :: (GameState, [Picture]) -> (GameState, [Picture])

Render the plane

renderEnemies :: (GameState, [Picture]) -> (GameState, [Picture])

Render enemies

renderScore :: (GameState, [Picture]) -> (GameState, [Picture])

Render the score

eventHandle :: Event -> GameState -> GameState

Handle keyboard input

data Direction

Because there is a Left/Right definition already

Constructors

DLeft

DRight

movePlane :: Direction -> GameState -> GameState

Move the plane horizontally

step :: Float -> GameState -> GameState

Combine the steps

stepSpeedup :: (Float, GameState) -> (Float, GameState)

Speedup game

stepPlane :: (Float, GameState) -> (Float, GameState)

Move plane according to pressed keys

mvPlane :: KeyState -> KeyState -> GameState -> GameState

Implementation of plane movement

filterEnemies :: (Float, GameState) -> (Float, GameState)

Remove out of screen enemies and increase score

addEnemy :: (Float, GameState) -> (Float, GameState)

Add next enemy according to difficulty index

stepEnemies :: (Float, GameState) -> (Float, GameState)

Move enemies in the direction of y axis

stepCollision :: (Float, GameState) -> (Float, GameState)

determine collisions between plane and enemies

distance :: Enemy -> (Float, Float) -> Float

Calculate the distance between two points

data GameState

Hold the state of the game

Constructors

Game

Fields

planeLoc :: Float

gameSpeed :: Float

enemies :: [Enemy]

randsrc :: [Float]

aState :: KeyState

dState :: KeyState

stopped :: Bool

score :: Int

difficulty :: Float

type Enemy = (Float, Float)

initialState :: GameState

The initial state of game

resetGame :: GameState -> GameState

Reset game state to initial values, except the random source

main :: IO ()

Main function